refactor: streamline model imports and enhance type checking
- Removed unused imports and consolidated model classes for Chai-1, Boltz-2, and ESMFold to improve code clarity and maintainability.
- Introduced type checking for core model classes to enhance code robustness.
- Updated import paths for core classes to reflect new module structure.
- Cleaned up commented-out code and unnecessary dependencies to optimize performance.
